+++ backend/estuary/api.go
@@ -2,15 +2,22 @@ package estuary
 
 import (
 	"context"
+	"errors"
 	"fmt"
+	"github.com/rclone/rclone/fs"
 	"github.com/rclone/rclone/lib/rest"
 	"net/http"
 	"net/url"
+	"time"
 )
 
 const (
 	colUuid = "coluuid"
 	colDir  = "dir"
+)
+
+const (
+	errNoUploadEndpoint = errors.New("No upload endpoint for object")
 )
 
 type CollectionCreate struct {
@@ -21,6 +28,32 @@ type CollectionCreate struct {
 type DeleteContentFromCollectionBody struct {
 	By    string `json:"by"`
 	Value string `json:"value"`
+}
+
+type ContentAdd struct {
+	ID      uint   `json:"estuaryId"`
+	Cid     string `json:"cid,omitempty"`
+	Error   string `json:"error"`
+	Details string `json:"details"`
+}
+
+type IpfsPin struct {
+	CID     string                 `json:"cid"`
+	Name    string                 `json:"name"`
+	Origins []string               `json:"origins"`
+	Meta    map[string]interface{} `json:"meta"`
+}
+
+type ContentByCID struct {
+	Content Content `json:"content"`
+}
+
+type Collection struct {
+	UUID        string    `json:"uuid"`
+	CreatedAt   time.Time `json:"createdAt"`
+	Name        string    `json:"name"`
+	Description string    `json:"description"`
+	UserID      uint      `json:"userId"`
 }
 
 func (f *Fs) fetchViewer(ctx context.Context) (response ViewerResponse, err error) {
@@ -132,5 +165,37 @@ func (o *Object) removeContentFromCollection(ctx context.Context, collectionId s
 	})
 
 	return err
+}
+
+func (o *Object) addContent(ctx context.Context, opts rest.Opts) (result ContentAdd, err error) {
+	endpoints := o.fs.viewer.Settings.UploadEndpoints
+
+	if len(endpoints) == 0 {
+		return ContentAdd{}, errNoUploadEndpoint
+	}
+
+	endpoint := 0
+
+	var response *http.Response
+	err = o.fs.pacer.Call(func() (bool, error) {
+		if endpoint == len(endpoints) {
+			return false, errAllEndpointsFailed
+		}
+
+		// Note: "Path" is actually embedded in the upload endpoint, which we use as the RootURL
+		opts.RootURL = endpoints[endpoint]
+		response, err = o.fs.client.CallJSON(ctx, &opts, nil, &result)
+		if contentAddingDisabled(response, err) {
+			fs.Debugf(o, "failed upload, retry w/ next upload endpoint")
+			endpoint += 1
+			return true, err
+		}
+
+		return shouldRetry(ctx, response, err)
+	})
+	return result, err
+}
+
+func (o *Object) replacePin(ctx context.Context) {
 
 }
